add inheritance asserv_default and asservExtended. add inheritance actions_default and actionsExtended. bug using several moveD...to be fixed.

+++ trunk-fusion/src/Common/Asserv/MotorControl.cpp
@@ -32,12 +32,12 @@ void MotorControl::stopMotors()
 }
 
 
-void MotorControl::setMotorLeftPosition(int tick, int power)
+void MotorControl::setMotorLeftPosition(long tick, int power)
 {
 	asservdriver->setMotorLeftPosition(tick, power);
 }
 
-void MotorControl::setMotorRightPosition(int tick, int power)
+void MotorControl::setMotorRightPosition(long tick, int power)
 {
 	asservdriver->setMotorRightPosition(tick, power);
 }
