+++ trunk-fusion/src/Driver.EV3/AsservDriver.hpp
@@ -41,8 +41,8 @@ public:
 	{
 	}
 
-	virtual void setMotorLeftPosition(int ticks, int power);
-	virtual void setMotorRightPosition(int ticks, int power);
+	virtual void setMotorLeftPosition(long ticks, int power);
+	virtual void setMotorRightPosition(long ticks, int power);
 
 	virtual void setMotorLeftPower(int power, int time);
 	virtual void setMotorRightPower(int power, int time);
